added sandbox selector to identity map data element view
