+++ extension.json
@@ -390,6 +390,74 @@
       "schema": {
         "$schema": "http://json-schema.org/draft-04/schema#",
         "type": "object",
+        "properties": {
+          "sandbox": {
+            "oneOf":[
+              {
+                "type": "string"
+              },
+              {
+                "type": "object",
+                "additionalProperties": {
+                  "type": "array",
+                  "items": {
+                    "type": "object",
+                    "properties": {
+                      "id": {
+                        "type": "string"
+                      },
+                      "authenticatedState": {
+                        "type": "string",
+                        "enum": [
+                          "loggedOut",
+                          "authenticated",
+                          "ambiguous"
+                        ]
+                      },
+                      "primary": {
+                        "type": "boolean"
+                      }
+                    },
+                    "additionalProperties": false
+                  }
+                }
+              }
+            ]
+          },
+          "items": {
+            "oneOf": [
+              {
+                "type": "string"
+              },
+              {
+                "type": "object",
+                "additionalProperties": {
+                  "type": "array",
+                  "items": {
+                    "type": "object",
+                    "properties": {
+                      "id": {
+                        "type": "string"
+                      },
+                      "authenticatedState": {
+                        "type": "string",
+                        "enum": [
+                          "loggedOut",
+                          "authenticated",
+                          "ambiguous"
+                        ]
+                      },
+                      "primary": {
+                        "type": "boolean"
+                      }
+                    },
+                    "additionalProperties": false
+                  }
+                }
+              }
+            ]
+          }
+        },
         "additionalProperties": {
           "type": "array",
           "items": {
